requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
Pillow
